Refactors groups into their own file
