+++ tools/tck-inspection/src/main/scala/org/opencypher/tools/tck/inspection/Group.scala
@@ -1,0 +1,53 @@
+/*
+ * Copyright (c) 2015-2020 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Attribution Notice under the terms of the Apache License 2.0
+ *
+ * This work was created by the collective efforts of the openCypher community.
+ * Without limiting the terms of Section 6, any Derivative Work that is not
+ * approved by the public consensus process of the openCypher Implementers Group
+ * should not be described as “Cypher” (and Cypher® is a registered trademark of
+ * Neo4j Inc.) or as "openCypher". Extensions by implementers or prototypes or
+ * proposals for change that have been documented or implemented should only be
+ * described as "implementation extensions to Cypher" or as "proposed changes to
+ * Cypher that are not yet approved by the openCypher community".
+ */
+package org.opencypher.tools.tck.inspection
+
+trait Group {
+  def name: String
+  def indent: Int
+  def parent: Option[Group]
+
+  override def toString: String = name
+}
+
+case object Total extends Group {
+  override val name = "Total"
+  override val indent = 0
+  override val parent: Option[Group] = None
+}
+
+case class Tag(override val name: String) extends Group {
+  override val indent = 1
+  override val parent: Option[Group] = Some(Total)
+}
+
+case class Feature(override val name: String, override val indent: Int, override val parent: Option[Group]) extends Group {
+  override def toString: String = "Feature: " + name
+}
+
+case class ScenarioCategory(override val name: String, override val indent: Int, override val parent: Option[Group]) extends Group
